refactor dialect/formatting to its own file
